feat: add facets
